+++ build-logic/src/main/groovy/org/omegat/module/OmegatModulePlugin.groovy
@@ -17,6 +17,8 @@ import org.gradle.api.tasks.compile.JavaCompile
 import org.gradle.api.tasks.testing.Test
 import org.gradle.process.ExecOperations
 
+import org.omegat.gradle.SignNativeJarTask
+
 import javax.inject.Inject
 import java.util.zip.ZipFile
